Move GuideText, GuideTextInvalid to shared components directory
