+++ src/modules/asset/components/TxInputViewer.tsx
@@ -3,8 +3,8 @@ import { View, Platform } from 'react-native';
 import { useTranslation } from 'react-i18next';
 import CryptoType from '../../../enums/CryptoType';
 import commaFormatter from '../../../utiles/commaFormatter';
-import GuideText from './GuideText';
-import GuideTextInvalid from './GuideTextInvalid';
+import GuideText from '../../../shared/components/GuideText';
+import GuideTextInvalid from '../../../shared/components/GuideTextInvalid';
 import LargeTextInput from './LargeTextInput';
 import AppColors from '../../../enums/AppColors';
 import PurposeType from '../../../enums/PurposeType';
