+++ src/shared/assets/images/alert_icon_xxhdpi.png
