+++ src/shared/components/GuideText.tsx
@@ -1,17 +1,14 @@
 import React from 'react';
 import { Text, StyleProp, TextStyle } from 'react-native';
-import AppColors from '../../../enums/AppColors';
-import AppFonts from '../../../enums/AppFonts';
+import AppColors from '../../enums/AppColors';
+import AppFonts from '../../enums/AppFonts';
 
 interface Props {
   text: string;
   style?: StyleProp<TextStyle>;
 }
 
-const GuideText: React.FC<Props> = ({
-  text,
-  style,
-}) => {
+const GuideText: React.FC<Props> = ({ text, style }) => {
   return (
     <Text
       style={{
@@ -20,8 +17,7 @@ const GuideText: React.FC<Props> = ({
         color: AppColors.SUB_BLACK,
         fontSize: 12,
         fontFamily: AppFonts.Regular,
-      }}
-    >
+      }}>
       {text}
     </Text>
   );
