+++ src/shared/components/GuideTextInvalid.tsx
@@ -1,17 +1,14 @@
 import React from 'react';
 import { View, Image, Text, StyleProp, ViewStyle } from 'react-native';
-import AppFonts from '../../../enums/AppFonts';
-import AppColors from '../../../enums/AppColors';
+import AppFonts from '../../enums/AppFonts';
+import AppColors from '../../enums/AppColors';
 
 interface Props {
   text: string;
   style?: StyleProp<ViewStyle>;
 }
 
-const GuideTextInvalid: React.FC<Props> = ({
-  text,
-  style,
-}) => {
+const GuideTextInvalid: React.FC<Props> = ({ text, style }) => {
   return (
     <View
       style={{
@@ -21,10 +18,9 @@ const GuideTextInvalid: React.FC<Props> = ({
         justifyContent: 'flex-end',
         alignItems: 'center',
         width: '100%',
-      }}
-    >
+      }}>
       <Image
-        source={require('../images/alert_icon_xxhdpi.png')}
+        source={require('../assets/images/alert_icon_xxhdpi.png')}
         style={{
           width: 15,
           height: 15,
@@ -37,8 +33,7 @@ const GuideTextInvalid: React.FC<Props> = ({
           color: AppColors.ERROR_RED,
           fontSize: 12,
           fontFamily: AppFonts.Medium,
-        }}
-      >
+        }}>
         {text}
       </Text>
     </View>
